Generate implementation of Parse trait when possible (closes #21)
Generate implementation of Parse trait when possible. The code now
generates 3 functions instead of one (parse):
- parse_be: parse object as big-endian
- parse_le: parse object as little-endian
- parse: default function, wraps a call to parse_be

If the endianness of the struct is fixed, then all 3 functions are
equivalent.

This helps generalizing the use of the Parse trait.

When there are extra args or a selector, it is not possible to generate
the trait implementation (function signatures are different). In that
case, an implementation block is generate with the 3 functions.

The code had to be refactored to generate several methods, and to be
able to handle some tricky corner cases, like generating code that
compiles, with trait generics/where clauses and function generics/where
clauses, that is accepted by the compiler.

Documentation is not yet updated.